Turns out, some bits (update program params) have to be pulsed in order to get the dispenser to actually accept updates.
After thinking about it a lot, I have decided that it is probably time to do total overhaul, since it is very awkward to send pulses with the current archetype.

Most of the code can be re-used, but I want to use the outgoing message queue to enforce the order in which messages are sent and received, and to only send a new message once the previous one has been acknowledged.  This is to avoid issues around the order of messages, since I will have to send lots of digital pulses.